+++ lib/commands/asm/websites/websitesclient._js
@@ -22,7 +22,7 @@ var __ = require('underscore');
 
 var profile = require('../../../util/profile');
 var utils = require('../../../util/utils');
-var validate = require('../../../util/validation');
+var validate = require('./validation');
 var cacheUtils = require('../../../util/cacheUtils');
 
 /*jshint camelcase:false*/
