+++ lib/util/validation.js
@@ -78,74 +78,3 @@ exports.parseDateTime = function (stringDateTime) {
     throw new Error($('The date format is incorrect'));
   }
 };
-
-// common functions for validating arguments
-
-function throwMissingArgument(name, func) {
-  throw new Error('Required argument ' + name + ' for function ' + func + ' is not defined');
-}
-
-function ArgumentValidator(functionName) {
-  this.func = functionName;
-}
-
-_.extend(ArgumentValidator.prototype, {
-  string: function (val, name) {
-    if (typeof val != 'string' || val.length === 0) {
-      throwMissingArgument(name, this.func);
-    }
-  },
-
-  object: function (val, name) {
-    if (!val) {
-      throwMissingArgument(name, this.func);
-    }
-  },
-
-  exists: function (val, name) {
-    this.object(val, name);
-  },
-
-  function: function (val, name) {
-    if (typeof val !== 'function') {
-      throw new Error('Parameter ' + name + ' for function ' + this.func + ' should be a function but is not');
-    }
-  },
-
-  value: function (val, name) {
-    if (!val) {
-      throwMissingArgument(name, this.func);
-    }
-  },
-
-  nonEmptyArray: function (val, name) {
-    if (!val || val.length === 0) {
-      throw new Error('Required array argument ' + name + ' for function ' + this.func + ' is either not defined or empty');
-    }
-  },
-
-  callback: function (val) {
-    this.object(val, 'callback');
-    this.function(val, 'callback');
-  },
-
-  test: function (predicate, message) {
-    if (!predicate()) {
-      throw new Error(message + ' in function ' + this.func);
-    }
-  },
-
-  tableNameIsValid: exports.tableNameIsValid,
-  containerNameIsValid: exports.containerNameIsValid,
-  blobNameIsValid: exports.blobNameIsValid,
-  pageRangesAreValid: exports.pageRangesAreValid,
-  queueNameIsValid: exports.queueNameIsValid
-});
-
-function validateArgs(functionName, validationRules) {
-  var validator = new ArgumentValidator(functionName);
-  validationRules(validator);
-}
-
-exports.ArgumentValidator = ArgumentValidator;
-exports.validateArgs = validateArgs;
